[revert] 5c2fc07 커밋을 되돌리는데 부모 트리를 952c64e 커밋으로 탐
This reverts commit 5c2fc0772ae9a046a93a3c3178fc05e4d9efb012, reversing
changes made to 952c64e7842dc0533da8bbbf2cad908ec6e1361c.
