--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/LocationPin/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
