--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/LocationPin/locationPinBlue.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "locationPinBlue.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
