--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/LocationPin/locationPinPink.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "locationPinPink.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
