--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/LocationPin/locationPinRed.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "locationPinRed.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
