--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/LocationPin/locationPinYellow.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "locationPinYellow.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
