--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignMeteor/warnSignMeteor3.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignMeteor3.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
