--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignMeteor/warnSignMeteor4.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignMeteor4.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
