--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignMeteor/warnSignMeteor5.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignMeteor5.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
