--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignMeteor/warnSignMeteor6.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignMeteor6.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
