--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignThunder/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
