--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignThunder/warnSignThunder2.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignThunder2.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
