--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignThunder/warnSignThunder3.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignThunder3.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
