--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignThunder/warnSignThunder5.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignThunder5.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
