--- DrivePalWidgetExtension/Assets.xcassets/CompactDynamicIsland/WarnSignThunder/warnSignThunder6.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "warnSignThunder6.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
